Integrate the ExistingInfraMachine controller from pkg/ to controllers/
