+++ controllers/cluster.weave.works/existinginframachine_controller_test.go
@@ -1,4 +1,4 @@
-package wks
+package controllers
 
 import (
 	"testing"
